refactor: #78 設定画面にタブを作って各種設定を設定に寄せる
- マスタアイテムの管理を設定画面の中にまとめるために設定画面をタブ化
- UserPreferenceの設定は2つのタブに分割（一般とGoogleカレンダー）
- アカウントも設定画面のタブにする
- タイムテーブルをタイムラインに変更
