+++ package.json
@@ -43,6 +43,7 @@
     "node-window-manager": "^2.2.4",
     "notistack": "^3.0.1",
     "react-big-calendar": "^1.8.1",
+    "react-color": "^2.19.3",
     "react-hook-form": "^7.45.0",
     "react-rnd": "^10.4.1",
     "react-router-dom": "^6.14.0",
@@ -58,6 +59,7 @@
     "@types/nedb": "^1.8.13",
     "@types/node": "^18.16.16",
     "@types/react": "^18.2.8",
+    "@types/react-color": "^3.0.6",
     "@types/react-dom": "^18.2.4",
     "@typescript-eslint/eslint-plugin": "^5.59.8",
     "@typescript-eslint/parser": "^5.59.8",
