--- src/renderer/src/pages/TimeTablePage.tsx
@@ -1,14 +1,0 @@
-import TimeTable from '@renderer/components/timeTable/TimeTable';
-
-const TimeTablePage = (): JSX.Element => {
-  console.log('TimeTablePage');
-
-  return (
-    <>
-      <p>タイムテーブル</p>
-      <TimeTable />
-    </>
-  );
-};
-
-export default TimeTablePage;
