+++ src/renderer/src/pages/TimelinePage.tsx
@@ -1,0 +1,6 @@
+import TimeTable from '@renderer/components/timeTable/TimeTable';
+
+export const TimelinePage = (): JSX.Element => {
+  console.log('TimelinePage');
+  return <TimeTable />;
+};
